+++ src/platformio.ts
@@ -8,6 +8,7 @@ export class PlatformIO {
 
     constructor() {
         this._pioTerminal = new PioTerminal();
+        this.createStatusBarItems();
     }
 
     public build(): void {
@@ -25,7 +26,32 @@ export class PlatformIO {
         AppInsightsClient.sendEvent("openSerialMonitor");
     }
 
+    public searchLibrary(): void {
+        let query = "CANCELED";
+        vscode.window.showInputBox({ prompt: `Search for PlatformIO Library` }).then((input: string) => {
+            if (input !== undefined) {
+                query = input;
+                this._pioTerminal.showAndRun(`platformio lib search ${query}`);
+            }
+            AppInsightsClient.sendEvent("searchLibrary", { query });
+        });
+    }
+
     public onDidCloseTerminal(closedTerminal: vscode.Terminal): void {
         this._pioTerminal.onDidCloseTerminal(closedTerminal);
+    }
+
+    private createStatusBarItems(): void {
+        let openSerialMonitorStatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
+        openSerialMonitorStatusBarItem.command = "platformio.openSerialMonitor";
+        openSerialMonitorStatusBarItem.text = "$(plug) Serial Monitor";
+        openSerialMonitorStatusBarItem.tooltip = "Open Serial Monitor";
+        openSerialMonitorStatusBarItem.show();
+
+        let searchLibraryStatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 0);
+        searchLibraryStatusBarItem.command = "platformio.searchLibrary";
+        searchLibraryStatusBarItem.text = "$(search) Library";
+        searchLibraryStatusBarItem.tooltip = "Search PlatformIO Library";
+        searchLibraryStatusBarItem.show();
     }
 }