+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.0.2 (2017-03-18)
+* Add Serial Monitor in Status Bar
+* Add Library Search in Status Bar
+
 ## 0.0.1 (2017-03-17)
 * Initial Release